move Configuration and Momentum to features
