+++ openpathsampling/__init__.py
@@ -53,8 +53,6 @@ from volume import (Volume, VolumeCombination, VolumeFactory, VoronoiVolume,
 from tools import empty_snapshot_from_openmm_topology, snapshot_from_pdb, \
     to_openmm_topology, trajectory_from_mdtraj
 
-from tools import simtk_units_from_md_snapshot
-
 from topology import ToyTopology, Topology, MDTrajTopology
 
 from toy_dynamics.toy_pes import Gaussian, HarmonicOscillator, LinearSlope, \
